requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
crispy-bootstrap5==0.7
Django==4.1.2
django-crispy-forms==1.14.0
django-versatileimagefield==2.2
gunicorn==20.1.0
idna==3.4
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
python-magic-debian-bin==0.4.25
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
